% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_numbers}
\alias{add_numbers}
\title{Add Two Numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{Numeric. First number to add.}
\item{y}{Numeric. Second number to add.}
}
\value{
The sum of x and y.
}
\description{
A simple function that adds two numbers together.
}
\examples{
add_numbers(2, 3)
add_numbers(10, -5)
} 