% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_instances}
\alias{find_instances}
\title{Find instances of de-novo Homer motifs}
\usage{
find_instances(pos_file, genome, output_dir = tempdir())
}
\arguments{
\item{pos_file}{<#> (Genomic Ranges object)}

\item{genome}{<#> (Installed Homer genome) or (path to FASTA)}

\item{output_dir}{Path to output dir for Homer analysis. Defaults to tempdir()}
}
\value{
A tidy dataframe of instances of motifs.
}
\description{
Find instances of de-novo Homer motifs in a GRanges using Homer's annotatePeaks.pl script.
}
\details{
Must point to a Homer output directory to find a 'homerMotifs.all.motifs' file. Saves a the raw output file from annotatePeaks.pl as 'motifInstances.tab" in the same folder.
}
\seealso{
\code{\link{call_homer}} \code{\link{tempdir}}
}
\author{
Malte Thodberg
}
