% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{load_zillow_directory}
\alias{load_zillow_directory}
\title{Load Zillow directory}
\usage{
load_zillow_directory(path = ".")
}
\arguments{
\item{path}{A character string specifying the path to find Zillow archives.}
}
\description{
\code{load_zillow_directory} returns a dataframe of zillow estimate archives.
}
\details{
Used after collecting data over time with home_estimate().
}
