% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhm.R
\name{logposterior}
\alias{logposterior}
\title{Log-posterior of a Bayesian Heating Model}
\usage{
logposterior(bhm)
}
\arguments{
\item{bhm}{a fitted model returned by a call to \code{bhm()}}
}
\value{
a function of the model's parameters (currently K, tb, DHW and sigma)
}
\description{
Provides the log-posterior of a heating model given the data, as a function
of the model's parameters.
}

