% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_homer}
\alias{parse_homer}
\title{Parse de-novo motifs from Homer}
\usage{
parse_homer(output_dir = tempdir())
}
\arguments{
\item{output_dir}{Path to output dir for Homer analysis. Defaults to tempdir()}
}
\value{
A list containing a dataframe of de-novo motif results and a list of corresponding PWMs.
}
\description{
Parse de-novo motifs from a Homer analysis from output_dir. Should not normally be called by the user.
}
\details{
Parses from the "homerResults/" folder and "homerMotifs.all.motifs" file of a Homer analysis and merges the results.
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{call_homer}} \code{\link{tempdir}}
}

