% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_instances}
\alias{parse_instances}
\title{Parse instances of de-novo motifs from Homer}
\usage{
parse_instances(output_dir = tempdir())
}
\arguments{
\item{output_dir}{Path to output dir for Homer analysis. Defaults to tempdir()}
}
\value{
A dataframe of the position of motif occurences.
}
\description{
Parse de-novo motifs instances from a Homer analysis from output_dir. Should not normally be called by the user.
}
\details{
Parses from the "homerResults/motifInstances.tab" file of a Homer analysis and merges the results.
}
\seealso{
\code{\link{call_homer}} \code{\link{find_instances}} \code{\link{tempdir}}
}
\author{
Malte Thodberg
}
