% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_known}
\alias{parse_known}
\title{Parse known motifs from Homer}
\usage{
parse_known()
}
\value{
A dataframe of known motif results.
}
\description{
Parse known motifs form a Homer analysis from tempdir(). Should not normally be called by the user.
}
\details{
Parses from the "knownResults.txt" file of a Homer analysis.
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{call_homer}} \code{\link{tempdir}}
}

