% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{call_homer}
\alias{call_homer}
\title{Motif Enrichment with Homer}
\usage{
call_homer(pos_file, genome, mask = NULL, bg = NULL, len = NULL,
  size = NULL, S = NULL, mis = NULL, norevopp = NULL, nomotif = NULL,
  rna = NULL, mset = NULL, basic = NULL, bits = NULL, nocheck = NULL,
  mcheck = NULL, noknown = NULL, mknown = NULL, p = NULL)
}
\arguments{
\item{pos_file}{See findMotifsGenome.pl}

\item{genome}{See findMotifsGenome.pl}

\item{mask}{See findMotifsGenome.pl}

\item{bg}{See findMotifsGenome.pl}

\item{len}{See findMotifsGenome.pl}

\item{size}{See findMotifsGenome.pl}

\item{S}{See findMotifsGenome.pl}

\item{mis}{See findMotifsGenome.pl}

\item{norevopp}{See findMotifsGenome.pl}

\item{nomotif}{See findMotifsGenome.pl}

\item{rna}{See findMotifsGenome.pl}

\item{mset}{See findMotifsGenome.pl}

\item{basic}{See findMotifsGenome.pl}

\item{bits}{See findMotifsGenome.pl}

\item{nocheck}{See findMotifsGenome.pl}

\item{mcheck}{See findMotifsGenome.pl}

\item{noknown}{See findMotifsGenome.pl}

\item{mknown}{See findMotifsGenome.pl}

\item{p}{See findMotifsGenome.pl}
}
\value{
List with output: command line used, and (if selected), known results, de-novo results and de-novo PWMs.
}
\description{
Call the findMotifsGenome.pl from Homer from inside R.
}
\details{
Simple R-wrapper for Homer's findMotifsGenome.pl. Saves all temporary files to tempdir().
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{GR_to_BED}} \code{\link{tempdir}}
}

