% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{GR_to_BED}
\alias{GR_to_BED}
\title{GRanges to BED}
\usage{
GR_to_BED(GR)
}
\arguments{
\item{GR}{GRanges-object.}
}
\value{
Path to file.
}
\description{
Saves an GRanges-object as a BED-file and returns the file location.
}
\details{
Saves a GRanges object as a BED-file. Names are set to increasing integers and scores are set to feature widths. The BED file is save to temp_dir().
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{tempdir}} \code{\link{tempfile}}
}

