% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{home_estimate}
\alias{home_estimate}
\title{Zillow home estimate}
\usage{
home_estimate(x, zws_path = "~/zillow_key.txt")
}
\arguments{
\item{x}{A character string representing a single Zillow ID}

\item{zws_id}{A character string representing a your unique Zillow Web
Services key.}
}
\value{
XML from Zillow Web Service.  Details of the XML are maitained by
Zillow here - https://www.zillow.com/howto/api/GetZestimate.htm.
}
\description{
\code{home_estimate} returns data for one or more homes listed on Zillow.
}
\details{
Future versions will include the ability to pass in more than one
  property.

Link to zillow web services where users obtain an account.

Future versions will have different output options.
}
