% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{plot_zillow_directory}
\alias{plot_zillow_directory}
\title{Plot Zillow archives}
\usage{
plot_zillow_directory(path = ".", strip_font_size = 17, nudge_x = 4,
  nudge_y = 250, value_size = 4)
}
\arguments{
\item{path}{A character string specifying the path to find Zillow archives.}
}
\value{
ggplot
}
\description{
\code{plot_zillow_directory} plots a timeseries sparkline for each home in your
   archive.
}
